services/fmi/requirements.txt
fmiopendata==0.4.3
asyncio==3.4.3
python-dotenv==1.0.1
